Extract type converter and join enums
