+++ lib/data/model/color_gradient.dart
@@ -1,6 +1,4 @@
-import 'dart:convert';
 import 'dart:ui';
-import 'package:drift/drift.dart';
 import '../local/database.dart' as db_provider;
 
 // Custom gradient representation for UI
@@ -35,20 +33,5 @@ class ColorGradient {
       colors: colors,
       createdAt: createdAt,
     );
-  }
-}
-
-class ColorListConverter extends TypeConverter<List<Color>, String> {
-  const ColorListConverter();
-
-  @override
-  List<Color> fromSql(String fromDb) {
-    final List<dynamic> jsonList = jsonDecode(fromDb);
-    return jsonList.map((value) => Color(value as int)).toList();
-  }
-
-  @override
-  String toSql(List<Color> value) {
-    return jsonEncode(value.map((color) => color.toARGB32()).toList());
   }
 }
