+++ lib/data/model/enums.dart
@@ -1,0 +1,2 @@
+enum LoanStatus { ongoing, completed, overdue, cancelled }
+enum LoanType { given, taken }
