+++ lib/data/model/type_converter/loan_status_converter.dart
@@ -1,9 +1,7 @@
 import 'package:drift/drift.dart';
+import 'package:rocket_pocket/data/model/enums.dart';
 
-enum LoanStatus { ongoing, completed, overdue, cancelled }
-
-// This converter is used to convert the LoanStatus enum to a string for storage in the database
-// and back to the enum when reading from the database.
+/// A type converter for converting [LoanStatus] enum to a string for database storage
 class LoanStatusConverter extends TypeConverter<LoanStatus, String> {
   const LoanStatusConverter();
